+++ e-self/app/(instructor)/dashboard/create-course/handleSubmit.ts
